Feat: from double to int

+++ tank_game_project_02148/src/main/java/dk/dtu/Player.java
@@ -155,20 +155,19 @@ public class Player {
         return hitbox;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public double getAngle() {
+    public int getAngle() {
         return angle;
     }
 
     public boolean getShot(){
         return shot;
-        
     }
 }
